requirements.txt
discord.py[voice]
jishaku~=1.18.2.188
blissops
psutil
aiohttp~=3.6.2
matplotlib~=3.2.2
scikit-image~=0.17.2
Wand~=0.6.1
numpy~=1.19.0